+++ front-end/src/style/addHistory.scss
@@ -7,8 +7,9 @@
 
 .add-history-title {
   padding: 5px;
-  font-size: $font-md;
-  font-weight: 500;
+  padding-bottom: 12px;
+  font-family: 'AppleSDGothicNeoR';
+  font-size: $font-lg;
 }
 
 .add-history-sub-title {
