+++ front-end/src/style/history.scss
@@ -9,9 +9,3 @@
     font-size: $font-md;
   }
 }
-
-.container-box-title-md {
-  font-family: 'AppleSDGothicNeoB';
-  font-size: $font-md;
-  margin-left: 1rem;
-}
